Group_3/multi_agent_ai_deep_researcher/requirements.txt
# Core LLM & Agent Framework
langchain>=0.1.0
langchain-community>=0.0.20
langgraph>=0.0.20
ollama>=0.1.0

# Vector Store & Embeddings
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2

# Web & Data
tavily-python>=0.2.0
beautifulsoup4>=4.12.0
requests>=2.31.0

# Configuration & Validation
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0

# Reliability & Async
tenacity>=8.2.3

# Web UI
streamlit>=1.28.0
streamlit-chat>=0.1.1

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Utilities
tqdm>=4.66.0

# Graph Orchestration & State Management
langgraph>=0.0.40
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Retry & Error Handling
tenacity>=8.2.0

# Web UI & Streaming
streamlit>=1.28.0
streamlit-chat>=0.1.1

# Structured Logging
python-json-logger>=2.0.7

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Utilities
python-dateutil>=2.8.2